+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## (Unreleased)
 
 ### Monument
+- (#73) `splice_style = "calls"` will no longer generate splices over part heads.
 - (#73) `splice_weight` is now applied to splices over part heads.
 - (#73) `splice_style = "calls"` now works for cyclic compositions.
 - (#71) Allow specifying an exact count with e.g. `count = 224` rather than